configmap: Move configmap package
This moves configmap package one level up in order to share it
with the config package introduced in a later commit.
